group same query related files into their own folder/namespace
