+++ src/SFA.DAS.ProviderRelationships.Web/Controllers/HealthCheckController.cs
@@ -5,7 +5,8 @@ using MediatR;
 using SFA.DAS.Authorization.EmployerFeatures;
 using SFA.DAS.Authorization.Mvc;
 using SFA.DAS.ProviderRelationships.Application.Commands.RunHealthCheck;
-using SFA.DAS.ProviderRelationships.Application.Queries;
+
+using SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck;
 using SFA.DAS.ProviderRelationships.Web.RouteValues.HealthCheck;
 using SFA.DAS.ProviderRelationships.Web.ViewModels.HealthCheck;
