+++ src/SFA.DAS.ProviderRelationships/Application/Queries/FindProviderToAdd/FindProviderToAddQueryResult.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.ProviderRelationships.Application.Queries
+﻿namespace SFA.DAS.ProviderRelationships.Application.Queries.FindProviderToAdd
 {
     public class FindProviderToAddQueryResult
     {
