+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAccountProvider/GetAccountProviderQuery.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace SFA.DAS.ProviderRelationships.Application.Queries
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProvider
 {
     public class GetAccountProviderQuery : IRequest<GetAccountProviderQueryResult>
     {
