+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAccountProvider/GetAccountProviderQueryResult.cs
@@ -1,6 +1,6 @@
 using SFA.DAS.ProviderRelationships.Dtos;
 
-namespace SFA.DAS.ProviderRelationships.Application.Queries
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProvider
 {
     public class GetAccountProviderQueryResult
     {
