+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAccountProviderLegalEntity/GetAccountProviderLegalEntityQuery.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace SFA.DAS.ProviderRelationships.Application.Queries
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviderLegalEntity
 {
     public class GetAccountProviderLegalEntityQuery : IRequest<GetAccountProviderLegalEntityQueryResult>
     {
