+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAccountProviderLegalEntity/GetAccountProviderLegalEntityQueryResult.cs
@@ -1,6 +1,6 @@
 using SFA.DAS.ProviderRelationships.Dtos;
 
-namespace SFA.DAS.ProviderRelationships.Application.Queries
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviderLegalEntity
 {
     public class GetAccountProviderLegalEntityQueryResult
     {
