+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAccountProviders/GetAccountProvidersQueryResult.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using SFA.DAS.ProviderRelationships.Dtos;
 
-namespace SFA.DAS.ProviderRelationships.Application.Queries
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviders
 {
     public class GetAccountProvidersQueryResult
     {
