+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAddedAccountProvider/GetAddedAccountProviderQuery.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace SFA.DAS.ProviderRelationships.Application.Queries
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAddedAccountProvider
 {
     public class GetAddedAccountProviderQuery : IRequest<GetAddedAccountProviderQueryResult>
     {
