+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAddedAccountProvider/GetAddedAccountProviderQueryResult.cs
@@ -1,6 +1,6 @@
 using SFA.DAS.ProviderRelationships.Dtos;
 
-namespace SFA.DAS.ProviderRelationships.Application.Queries
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAddedAccountProvider
 {
     public class GetAddedAccountProviderQueryResult
     {
