+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetHealthCheck/GetHealthCheckQuery.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace SFA.DAS.ProviderRelationships.Application.Queries
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck
 {
     public class GetHealthCheckQuery : IRequest<GetHealthCheckQueryResult>
     {
