+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetProviderToAdd/GetProviderToAddQueryHandler.cs
@@ -9,7 +9,7 @@ using Microsoft.EntityFrameworkCore;
 using SFA.DAS.ProviderRelationships.Data;
 using SFA.DAS.ProviderRelationships.Dtos;
 
-namespace SFA.DAS.ProviderRelationships.Application.Queries
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetProviderToAdd
 {
     public class GetProviderToAddQueryHandler : IRequestHandler<GetProviderToAddQuery, GetProviderToAddQueryResult>
     {
